requirements.txt
beautifulsoup4==4.12.2
lxml
requests==2.31.0
selenium==4.15.0
webdriver-manager
python-dotenv==1.0.0
